Assigning location
Instead of hardcoded locations in the constuctor of the Google maps method, I have assigned the location to a variablem and I have also created a method that returns the location, giving us more control of the location function. In the future we can pass any values we want to the longitude and latitude variables.